Massive refactor of all things
This adds and deletes a whole whack of code. Should this be multiple
commits? Yes. Do I feel like breaking it up? No.

I can now generate games, look at the current room, and exit the room
by providing the exit id. When exiting the room (and in the future when
doing all actions) I provide the events that happened, and the actions
that can be taken inside of the current room.

I don't have a way to set a "current" game for a user, as I think they
can just always provide the game id. They will have to set a "player" as
their current, which lets people create and change the characters they
are using whenever they like, which I think is a fun idea.

I'm removing some room generation stuff because I think I think it was
cluttering up my API for now, where I want it to be focused on games.
Maybe I'll add it back in later.