Add start of Underworld Server docs
I'm going to use mdbook to generate the documentation for my game.

Right now the only exploration is done via the Open API, and that's
not exactly the easiest way to find what the actual flow should be in
the game.

I think some basic docs will go a long way towards me being able to
actually try to get this out there for people to use.