Rework args and how ids are taken in
Now the username is passed via the header. I'm doing this because it's
closer to what will be happening once I finally get some form of auth
into this.

Also moving the game id into the URL instead of the args, which means
I can now use the actions directly for the args.