Add security middleware and schema
This lets us lock down parts of our API, but not all of it. The
middleware is to make it so the frontend (for the most part) doesn't
need to pass along the API token, if it had been authed from the
backend then the cookie will be passed directly and that will get
put into the header.